Refactor away add volunteer form
