+++ src/sections/admin/destinations/destination/index.jsx
@@ -28,10 +28,6 @@ class Destination extends Component {
                     uri: `/admin/destinations/${this.props.params.destinationId}/coordinators`
                 },
                 {
-                    name: 'Add coordinator',
-                    uri: `/admin/destinations/${this.props.params.destinationId}/coordinators/new`
-                },
-                {
                     name: 'Edit destination',
                     uri: `/admin/destinations/${this.props.params.destinationId}/edit`
                 },
